delete mythtv/libs/libmythtv/format.h
moving the used items to where they are used.
