+++ mythtv/libs/libmythtv/captions/cc708decoder.h
@@ -10,8 +10,6 @@
 #include "libmythbase/compat.h"
 #include "libmythbase/mythchrono.h"
 
-#include "format.h"
-
 using cc708_seen_flags = std::array<bool,64>;
 using cc708_seen_times = std::array<SystemTime,64>;
